Sketched event handler for Lity popup and D3 chart
Sketched the event hander function for opening the Lity lightbox.
When the <a> element is pressed to activate the lightbox,
traverse up the DOM from the <a> to find the Id number of
the single card. The Id will be used to get the stressor object
and associated heart rates from the STATE_DATA (or the DOM).
The object will be passed to the drawChart() and resizeChart()
functions to draw the D3 chart inside the Lity lightbox.